% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_categorical}
\alias{plot_categorical}
\title{Plot a categorical variable N and label with percent}
\usage{
plot_categorical(d, var, group = NULL, panel = NULL, stack = FALSE,
  margin = "no", yperc = FALSE, ylab = "Frequency")
}
\description{
Goal: have an example with stack=FALSE vs TRUE, margin = 'no', 'yes', or 'only'
I think I should externalize the tabulation - that would be better for unit
testing. Then the plot function would only need to add the label.
}
\details{
Goal: be able to add a group and/or panel variable using standard ggplot syntax
}
\examples{
# Simple example with the nsfg edgelist - cut length into only 2 categories
# This is with the y-scale being N and no stacking
library(egonet)
data(nsfg1)
dfl <- reshape_edgelist(nsfgs, delete_empty=c("active","once","rel","dfs","dls","len"))
d <- transform(dfl, lenf=cut(len,breaks=c(0,12,351)))
plot_cat(d, 'lenf', NULL, NULL)
plot_cat(d, 'lenf', 'sex', NULL)
plot_cat(d, 'lenf', 'sex', 'main') + facet_grid(.~panel)
# Now there is stacking, with y-scale still N (note that coord is flipped)
# Def need to add an option to turn off labels and turn on legend instead
plot_cat(d, 'lenf', stack=TRUE)
plot_cat(d, 'lenf', 'sex', stack=TRUE)
plot_cat(d, 'lenf', 'sex', 'main', stack=TRUE) + facet_grid(panel~.)
plot_cat(d, 'lenf', 'sex', 'main', yperc=TRUE, ylab='Percent') +
    facet_grid(panel~.)
plot_cat(pop, 'lencat1', 'sex', 'main', yperc=TRUE, ylab='Percent') +
    facet_grid(panel~.)
}

