% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undx.R
\name{setup_hivbackcalc}
\alias{setup_hivbackcalc}
\title{Setup function}
\usage{
setup_hivbackcalc(workd, datafile, source_these, loadlib = TRUE,
  msm = FALSE, na = "", load_package = TRUE, package_updated = NULL,
  packagefile = NULL)
}
\arguments{
\item{workd}{Working directory full file path}

\item{datafile}{Path to data, either full or from the workd}

\item{source_these}{Vector of paths to other files to source, either full or from the workd}

\item{loadlib}{Logical indicating whether to load libries}

\item{msm}{Logical indicating whether this is the KC MSM analysis where the data frame was called msm}

\item{na}{Character indicating whether NAs are "NA" or spaces ("")}

\item{load_package}{Logical indicating whether the HIVBackCalc package 
should be loaded (TRUE). If you want to have some package functions but overwrite others,
specify TRUE here and a file in the 'packagefile' param}

\item{package_updated}{DEPRECATED}

\item{packagefile}{Within the workd, path to file with newer package functions}
}
\description{
Setup function
}

